Update image.md
I made some minor edits that you will agree with, but I wonder if this headline "Hovered, Active and Disabled States" is a little misleading. The properties in the article are PressedUrl, HoveredUrl and DisabledUrl and I don't think "pressed" and "active" are the same thing. In example 2 you also discuss an image for the button in normal state. Why isn't "normal" part of the same headline? Is there also a property for normal that isn't discussed in the article?
Also, your slug for this article doesn't match the properties discussed in the article, so that's worse for search.